+++ cpplint_unittest.py
@@ -312,6 +312,14 @@ class TestCpplint(CpplintTestBase):
             "Do not indent within a namespace.  [whitespace/indent_namespace] [4]",
         ]
 
+        # In a namespace, but not *directly*
+        lines = [
+            "namespace Lumi {",
+            "class Outie {",
+            "    class Innie {",
+        ]
+        assert self.GetNamespaceResults(lines) == ""
+
     def testNamespaceIndentationIndentedParameter(self):
         lines = [
             "namespace Test {",
@@ -2970,6 +2978,14 @@ class TestCpplint(CpplintTestBase):
             "",
         )
 
+    def testInlineControlBodyWithLambda(self):
+        self.TestLint("if (all_of(v, [](auto b) { return b == 0; })) {", "")
+        self.TestLint(
+            "if (condition) { return true; }",
+            "Controlled statements inside brackets of if clause should be on a separate line"
+            "  [whitespace/newline] [5]",
+        )
+
     def testMismatchingSpacesInParens(self):
         self.TestLint("if (foo ) {", "Mismatching spaces inside () in if  [whitespace/parens] [5]")
         self.TestLint(
@@ -3775,6 +3791,18 @@ class TestCpplint(CpplintTestBase):
         self.TestBlankLinesCheck(["int x(\n", "    int a) {\n", "\n", "return 0;\n", "}"], 0, 0)
         self.TestBlankLinesCheck(
             ["int x(\n", "    int a) const {\n", "\n", "return 0;\n", "}"], 0, 0
+        )
+        self.TestBlankLinesCheck(
+            [
+                "Foo::Foo()\n",
+                "    : bar(),\n",
+                "      qi() {\n",
+                "\n",
+                "  DoSomething();\n",
+                "}",
+            ],
+            0,
+            0,
         )
         self.TestBlankLinesCheck(["int x(\n", "     int a) {\n", "\n", "return 0;\n", "}"], 1, 0)
         self.TestBlankLinesCheck(["int x(\n", "   int a) {\n", "\n", "return 0;\n", "}"], 1, 0)
@@ -5810,6 +5838,10 @@ func2();""",
         self.TestLint(
             '#include "bar.hh"',
             "Include the directory when naming header files  [build/include_subdir] [4]",
+        )
+        self.TestLint(
+            '#include "package/impl.c++"',
+            "Do not include .c++ files from other packages  [build/include] [4]",
         )
         self.TestLint('#include "baz.aa"', "")
         self.TestLint('#include "dir/foo.h"', "")